Fixed an issue where the countdown timer would not count down in the launch detail view.
Since updating the order of the detail cells the countdown timer would remain static at the value it appeared with. This was because the update function referenced cell 0 instead of cell 1.
It has been updated to use the countdown row value from the new detailSections enum hopefully preventing this issue from reoccurring should the order ever change again or if new sections are added.